Move bindingAdapters to library
