+++ sample/src/main/java/com/example/android_mvvm/sample/ExampleApplication.java
@@ -2,7 +2,11 @@ package com.example.android_mvvm.sample;
 
 import android.app.Application;
 import android.content.Context;
+import android.databinding.ViewDataBinding;
 
+import com.example.android_mvvm.ViewModel;
+import com.example.android_mvvm.adapters.ViewModelBinder;
+import com.example.android_mvvm.utils.BindingUtils;
 import com.squareup.leakcanary.LeakCanary;
 import com.squareup.leakcanary.RefWatcher;
 
@@ -18,5 +22,11 @@ public class ExampleApplication extends Application {
     public void onCreate() {
         super.onCreate();
         refWatcher = LeakCanary.install(this);
+        BindingUtils.setDefaultBinder(new ViewModelBinder() {
+            @Override
+            public void bind(ViewDataBinding viewDataBinding, ViewModel viewModel) {
+                viewDataBinding.setVariable(com.example.android_mvvm.sample.BR.vm, viewModel);
+            }
+        });
     }
 }
